Create very rough layout for container grid
Set up our initial element containers in the rough area where they will
eventually be placed. Eventually, the grid rows will be changed to
accomodate the  content placed within to be more responsive. .wrapper
has been set to flex so that we can place our content in the center of
the screen.